+++ tidy3d/components/simulation.py
@@ -1464,10 +1464,25 @@ class AbstractYeeGridSimulation(AbstractSimulation, ABC):
         if remove_outside_structures:
             new_structures = [strc for strc in self.structures if new_box.intersects(strc.geometry)]
         else:
-            new_structures = self.structures
+            new_structures = list(self.structures)
 
         if sources is None:
             sources = [src for src in self.sources if new_box.intersects(src)]
+
+        # some nonlinear materials depend on the central frequency
+        # we update them with hardcoded freq0
+        freqs = np.array([source.source_time.freq0 for source in self.sources])
+        for i, structure in enumerate(new_structures):
+            medium = structure.medium
+            nonlinear_spec = medium.nonlinear_spec
+            if nonlinear_spec is not None:
+                new_nonlinear_spec = nonlinear_spec._hardcode_medium_freqs(
+                    medium=medium, freqs=freqs
+                )
+                new_structure = structure.updated_copy(
+                    nonlinear_spec=new_nonlinear_spec, path="medium"
+                )
+                new_structures[i] = new_structure
 
         if monitors is None:
             monitors = [mnt for mnt in self.monitors if new_box.intersects(mnt)]